refactor(headless): prune server dead symbols
